asset_importer/wc1_xml2json/requirements.txt
fire
pillow
realesrgan
torch
torchvision
